refactor: split algo/utils.py into focused modules
pr_agent/algo/utils.py was a 1,500-line grab-bag. Three cohesive
clusters moved out, with utils re-exporting every moved name so all
existing 'from pr_agent.algo.utils import ...' call sites keep working:

- markdown_render.py: convert_to_markdown_v2 and its render helpers
  (emphasize_header, parse_code_suggestion, todo formatting, ...)
- output_repair.py: load_yaml/try_fix_yaml/try_fix_json repair helpers
  for malformed model output
- types.py (existing): Range, ModelType, TodoItem, PRReviewHeader,
  ReasoningEffort, PRDescriptionHeader join FilePatchInfo/EDIT_TYPE

Code is moved verbatim; no behavior change.
