+++ pr_agent/algo/types.py
@@ -1,6 +1,8 @@
 from dataclasses import dataclass
 from enum import Enum
-from typing import Optional
+from typing import Optional, Tuple, TypedDict
+
+from pydantic import BaseModel
 
 
 class EDIT_TYPE(Enum):
@@ -24,3 +26,40 @@ class FilePatchInfo:
     num_minus_lines: int = -1
     language: Optional[str] = None
     ai_file_summary: str = None
+class Range(BaseModel):
+    line_start: int  # should be 0-indexed
+    line_end: int
+    column_start: int = -1
+    column_end: int = -1
+
+
+class ModelType(str, Enum):
+    REGULAR = "regular"
+    WEAK = "weak"
+    REASONING = "reasoning"
+
+
+class TodoItem(TypedDict):
+    relevant_file: str
+    line_range: Tuple[int, int]
+    content: str
+
+
+class PRReviewHeader(str, Enum):
+    REGULAR = "## PR Reviewer Guide"
+    INCREMENTAL = "## Incremental PR Reviewer Guide"
+
+
+class ReasoningEffort(str, Enum):
+    XHIGH = "xhigh"
+    HIGH = "high"
+    MEDIUM = "medium"
+    LOW = "low"
+    MINIMAL = "minimal"
+    NONE = "none"
+
+
+class PRDescriptionHeader(str, Enum):
+    DIAGRAM_WALKTHROUGH = "Diagram Walkthrough"
+    FILE_WALKTHROUGH = "File Walkthrough"
+
