+++ pr_agent/algo/utils.py
@@ -1,32 +1,35 @@
 from __future__ import annotations
 
-import ast
-import copy
 import difflib
 import hashlib
-import html
 import json
 import os
 import re
 import sys
-import textwrap
 import time
 import traceback
 from datetime import datetime
-from enum import Enum
 from importlib.metadata import PackageNotFoundError, version
-from typing import Any, List, Optional, Tuple, TypedDict
+from typing import Any, List, Optional, Tuple
 
 import html2text
 import requests
 import yaml
-from pydantic import BaseModel
 from starlette_context import context
 
 from pr_agent.algo import MAX_TOKENS
-from pr_agent.algo.git_patch_processing import extract_hunk_lines_from_patch
+# Markdown rendering, output repair, and shared types moved to sibling modules;
+# re-exported here so existing `from pr_agent.algo.utils import ...` keeps working.
+from pr_agent.algo.markdown_render import (  # noqa: F401
+    convert_to_markdown_v2, emphasize_header, extract_relevant_lines_str,
+    format_todo_item, format_todo_items, is_value_no, parse_code_suggestion,
+    process_can_be_split, replace_code_tags, ticket_markdown_logic)
+from pr_agent.algo.output_repair import (fix_json_escape_char,  # noqa: F401
+                                         load_yaml, try_fix_json, try_fix_yaml)
 from pr_agent.algo.token_handler import TokenEncoder
-from pr_agent.algo.types import FilePatchInfo
+from pr_agent.algo.types import (FilePatchInfo, ModelType,  # noqa: F401
+                                 PRDescriptionHeader, PRReviewHeader, Range,
+                                 ReasoningEffort, TodoItem)
 from pr_agent.config_loader import get_settings, global_settings
 from pr_agent.log import get_logger
 
@@ -39,78 +42,12 @@ def get_model(model_type: str = "model_weak") -> str:
     return get_settings().config.model
 
 
-class Range(BaseModel):
-    line_start: int  # should be 0-indexed
-    line_end: int
-    column_start: int = -1
-    column_end: int = -1
-
-
-class ModelType(str, Enum):
-    REGULAR = "regular"
-    WEAK = "weak"
-    REASONING = "reasoning"
-
-
-class TodoItem(TypedDict):
-    relevant_file: str
-    line_range: Tuple[int, int]
-    content: str
-
-
-class PRReviewHeader(str, Enum):
-    REGULAR = "## PR Reviewer Guide"
-    INCREMENTAL = "## Incremental PR Reviewer Guide"
-
-
-class ReasoningEffort(str, Enum):
-    XHIGH = "xhigh"
-    HIGH = "high"
-    MEDIUM = "medium"
-    LOW = "low"
-    MINIMAL = "minimal"
-    NONE = "none"
-
-
-class PRDescriptionHeader(str, Enum):
-    DIAGRAM_WALKTHROUGH = "Diagram Walkthrough"
-    FILE_WALKTHROUGH = "File Walkthrough"
-
-
 def get_setting(key: str) -> Any:
     try:
         key = key.upper()
         return context.get("settings", global_settings).get(key, global_settings.get(key, None))
     except Exception:
         return global_settings.get(key, None)
-
-
-def emphasize_header(text: str, only_markdown=False, reference_link=None) -> str:
-    try:
-        # Finding the position of the first occurrence of ": "
-        colon_position = text.find(": ")
-
-        # Splitting the string and wrapping the first part in <strong> tags
-        if colon_position != -1:
-            # Everything before the colon (inclusive) is wrapped in <strong> tags
-            if only_markdown:
-                if reference_link:
-                    transformed_string = f"[**{text[:colon_position + 1]}**]({reference_link})\n" + text[colon_position + 1:]
-                else:
-                    transformed_string = f"**{text[:colon_position + 1]}**\n" + text[colon_position + 1:]
-            else:
-                if reference_link:
-                    transformed_string = f"<strong><a href='{reference_link}'>{text[:colon_position + 1]}</a></strong><br>" + text[colon_position + 1:]
-                else:
-                    transformed_string = "<strong>" + text[:colon_position + 1] + "</strong>" +'<br>' + text[colon_position + 1:]
-        else:
-            # If there's no ": ", return the original string
-            transformed_string = text
-
-        return transformed_string
-    except Exception as e:
-        get_logger().exception(f"Failed to emphasize header: {e}")
-        return text
 
 
 def unique_strings(input_list: List[str]) -> List[str]:
@@ -123,544 +60,6 @@ def unique_strings(input_list: List[str]) -> List[str]:
             unique_list.append(item)
             seen.add(item)
     return unique_list
-
-
-def convert_to_markdown_v2(output_data: dict,
-                           gfm_supported: bool = True,
-                           incremental_review=None,
-                           git_provider=None,
-                           files=None) -> str:
-    """
-    Convert a dictionary of data into markdown format.
-    Args:
-        output_data (dict): A dictionary containing data to be converted to markdown format.
-    Returns:
-        str: The markdown formatted text generated from the input dictionary.
-    """
-
-    emojis = {
-        "Can be split": "🔀",
-        "Key issues to review": "⚡",
-        "Recommended focus areas for review": "⚡",
-        "Score": "🏅",
-        "Relevant tests": "🧪",
-        "Focused PR": "✨",
-        "Relevant ticket": "🎫",
-        "Security concerns": "🔒",
-        "Todo sections": "📝",
-        "Insights from user's answers": "📝",
-        "Code feedback": "🤖",
-        "Estimated effort to review [1-5]": "⏱️",
-        "Contribution time cost estimate": "⏳",
-        "Ticket compliance check": "🎫",
-    }
-    markdown_text = ""
-    if not incremental_review:
-        markdown_text += f"{PRReviewHeader.REGULAR.value} 🔍\n\n"
-    else:
-        markdown_text += f"{PRReviewHeader.INCREMENTAL.value} 🔍\n\n"
-        markdown_text += f"⏮️ Review for commits since previous PR-Agent review {incremental_review}.\n\n"
-    if not output_data or not output_data.get('review', {}):
-        return ""
-
-    if get_settings().get("pr_reviewer.enable_intro_text", False):
-        markdown_text += f"Here are some key observations to aid the review process:\n\n"
-
-    if gfm_supported:
-        markdown_text += "<table>\n"
-
-    todo_summary = output_data['review'].pop('todo_summary', '')
-    for key, value in output_data['review'].items():
-        if value is None or value == '' or value == {} or value == []:
-            if key.lower() not in ['can_be_split', 'key_issues_to_review']:
-                continue
-        key_nice = key.replace('_', ' ').capitalize()
-        emoji = emojis.get(key_nice, "")
-        if 'Estimated effort to review' in key_nice:
-            key_nice = 'Estimated effort to review'
-            value = str(value).strip()
-            if value.isnumeric():
-                value_int = int(value)
-            else:
-                try:
-                    value_int = int(value.split(',')[0])
-                except ValueError:
-                    continue
-            blue_bars = '🔵' * value_int
-            white_bars = '⚪' * (5 - value_int)
-            value = f"{value_int} {blue_bars}{white_bars}"
-            if gfm_supported:
-                markdown_text += f"<tr><td>"
-                markdown_text += f"{emoji}&nbsp;<strong>{key_nice}</strong>: {value}"
-                markdown_text += f"</td></tr>\n"
-            else:
-                markdown_text += f"### {emoji} {key_nice}: {value}\n\n"
-        elif 'relevant tests' in key_nice.lower():
-            value = str(value).strip().lower()
-            if gfm_supported:
-                markdown_text += f"<tr><td>"
-                if is_value_no(value):
-                    markdown_text += f"{emoji}&nbsp;<strong>No relevant tests</strong>"
-                else:
-                    markdown_text += f"{emoji}&nbsp;<strong>PR contains tests</strong>"
-                markdown_text += f"</td></tr>\n"
-            else:
-                if is_value_no(value):
-                    markdown_text += f'### {emoji} No relevant tests\n\n'
-                else:
-                    markdown_text += f"### {emoji} PR contains tests\n\n"
-        elif 'ticket compliance check' in key_nice.lower():
-            markdown_text = ticket_markdown_logic(emoji, markdown_text, value, gfm_supported)
-        elif 'contribution time cost estimate' in key_nice.lower():
-            if gfm_supported:
-                markdown_text += f"<tr><td>{emoji}&nbsp;<strong>Contribution time estimate</strong> (best, average, worst case): "
-                markdown_text += f"{value['best_case'].replace('m', ' minutes')} | {value['average_case'].replace('m', ' minutes')} | {value['worst_case'].replace('m', ' minutes')}"
-                markdown_text += f"</td></tr>\n"
-            else:
-                markdown_text += f"### {emoji} Contribution time estimate (best, average, worst case): "
-                markdown_text += f"{value['best_case'].replace('m', ' minutes')} | {value['average_case'].replace('m', ' minutes')} | {value['worst_case'].replace('m', ' minutes')}\n\n"
-        elif 'security concerns' in key_nice.lower():
-            if gfm_supported:
-                markdown_text += f"<tr><td>"
-                if is_value_no(value):
-                    markdown_text += f"{emoji}&nbsp;<strong>No security concerns identified</strong>"
-                else:
-                    markdown_text += f"{emoji}&nbsp;<strong>Security concerns</strong><br><br>\n\n"
-                    value = emphasize_header(value.strip())
-                    markdown_text += f"{value}"
-                markdown_text += f"</td></tr>\n"
-            else:
-                if is_value_no(value):
-                    markdown_text += f'### {emoji} No security concerns identified\n\n'
-                else:
-                    markdown_text += f"### {emoji} Security concerns\n\n"
-                    value = emphasize_header(value.strip(), only_markdown=True)
-                    markdown_text += f"{value}\n\n"
-        elif 'todo sections' in key_nice.lower():
-            if gfm_supported:
-                markdown_text += "<tr><td>"
-                if is_value_no(value):
-                    markdown_text += f"✅&nbsp;<strong>No TODO sections</strong>"
-                else:
-                    markdown_todo_items = format_todo_items(value, git_provider, gfm_supported)
-                    markdown_text += f"{emoji}&nbsp;<strong>TODO sections</strong>\n<br><br>\n"
-                    markdown_text += markdown_todo_items
-                markdown_text += "</td></tr>\n"
-            else:
-                if is_value_no(value):
-                    markdown_text += f"### ✅ No TODO sections\n\n"
-                else:
-                    markdown_todo_items = format_todo_items(value, git_provider, gfm_supported)
-                    markdown_text += f"### {emoji} TODO sections\n\n"
-                    markdown_text += markdown_todo_items
-        elif 'can be split' in key_nice.lower():
-            if gfm_supported:
-                markdown_text += f"<tr><td>"
-                markdown_text += process_can_be_split(emoji, value)
-                markdown_text += f"</td></tr>\n"
-        elif 'key issues to review' in key_nice.lower():
-            # value is a list of issues
-            if is_value_no(value):
-                if gfm_supported:
-                    markdown_text += f"<tr><td>"
-                    markdown_text += f"{emoji}&nbsp;<strong>No major issues detected</strong>"
-                    markdown_text += f"</td></tr>\n"
-                else:
-                    markdown_text += f"### {emoji} No major issues detected\n\n"
-            else:
-                issues = value
-                if gfm_supported:
-                    markdown_text += f"<tr><td>"
-                    # markdown_text += f"{emoji}&nbsp;<strong>{key_nice}</strong><br><br>\n\n"
-                    markdown_text += f"{emoji}&nbsp;<strong>Recommended focus areas for review</strong><br><br>\n\n"
-                else:
-                    markdown_text += f"### {emoji} Recommended focus areas for review\n\n#### \n"
-                for i, issue in enumerate(issues):
-                    try:
-                        if not issue or not isinstance(issue, dict):
-                            continue
-                        relevant_file = issue.get('relevant_file', '').strip()
-                        issue_header = issue.get('issue_header', '').strip()
-                        if issue_header.lower() == 'possible bug':
-                            issue_header = 'Possible Issue'  # Make the header less frightening
-                        issue_content = issue.get('issue_content', '').strip()
-                        start_line = int(str(issue.get('start_line', 0)).strip())
-                        end_line = int(str(issue.get('end_line', 0)).strip())
-
-                        relevant_lines_str = extract_relevant_lines_str(end_line, files, relevant_file, start_line, dedent=True)
-                        if git_provider:
-                            reference_link = git_provider.get_line_link(relevant_file, start_line, end_line)
-                        else:
-                            reference_link = None
-
-                        if gfm_supported:
-                            if reference_link is not None and len(reference_link) > 0:
-                                if relevant_lines_str:
-                                    issue_str = f"<details><summary><a href='{reference_link}'><strong>{issue_header}</strong></a>\n\n{issue_content}\n</summary>\n\n{relevant_lines_str}\n\n</details>"
-                                else:
-                                    issue_str = f"<a href='{reference_link}'><strong>{issue_header}</strong></a><br>{issue_content}"
-                            else:
-                                issue_str = f"<strong>{issue_header}</strong><br>{issue_content}"
-                        else:
-                            if reference_link is not None and len(reference_link) > 0:
-                                issue_str = f"[**{issue_header}**]({reference_link})\n\n{issue_content}\n\n"
-                            else:
-                                issue_str = f"**{issue_header}**\n\n{issue_content}\n\n"
-                        markdown_text += f"{issue_str}\n\n"
-                    except Exception as e:
-                        get_logger().exception(f"Failed to process 'Recommended focus areas for review': {e}")
-                if gfm_supported:
-                    markdown_text += f"</td></tr>\n"
-        else:
-            if gfm_supported:
-                markdown_text += f"<tr><td>"
-                markdown_text += f"{emoji}&nbsp;<strong>{key_nice}</strong>: {value}"
-                markdown_text += f"</td></tr>\n"
-            else:
-                markdown_text += f"### {emoji} {key_nice}: {value}\n\n"
-
-    if gfm_supported:
-        markdown_text += "</table>\n"
-
-    return markdown_text
-
-
-def extract_relevant_lines_str(end_line, files, relevant_file, start_line, dedent=False) -> str:
-    """
-    Finds 'relevant_file' in 'files', and extracts the lines from 'start_line' to 'end_line' string from the file content.
-    """
-    try:
-        relevant_lines_str = ""
-        if files:
-            files = set_file_languages(files)
-            for file in files:
-                if file.filename.strip() == relevant_file:
-                    if not file.head_file:
-                        # as a fallback, extract relevant lines directly from patch
-                        patch = file.patch
-                        get_logger().info(f"No content found in file: '{file.filename}' for 'extract_relevant_lines_str'. Using patch instead")
-                        _, selected_lines = extract_hunk_lines_from_patch(patch, file.filename, start_line, end_line,side='right')
-                        if not selected_lines:
-                            get_logger().error(f"Failed to extract relevant lines from patch: {file.filename}")
-                            return ""
-                        # filter out '-' lines
-                        relevant_lines_str = ""
-                        for line in selected_lines.splitlines():
-                            if line.startswith('-'):
-                                continue
-                            relevant_lines_str += line[1:] + '\n'
-                    else:
-                        relevant_file_lines = file.head_file.splitlines()
-                        relevant_lines_str = "\n".join(relevant_file_lines[start_line - 1:end_line])
-
-                    if dedent and relevant_lines_str:
-                        # Remove the longest leading string of spaces and tabs common to all lines.
-                        relevant_lines_str = textwrap.dedent(relevant_lines_str)
-                    relevant_lines_str = f"```{file.language}\n{relevant_lines_str}\n```"
-                    break
-
-        return relevant_lines_str
-    except Exception as e:
-        get_logger().exception(f"Failed to extract relevant lines: {e}")
-        return ""
-
-
-def ticket_markdown_logic(emoji, markdown_text, value, gfm_supported) -> str:
-    ticket_compliance_str = ""
-    compliance_emoji = ''
-    # Track compliance levels across all tickets
-    all_compliance_levels = []
-
-    if isinstance(value, list):
-        for ticket_analysis in value:
-            try:
-                ticket_url = ticket_analysis.get('ticket_url', '').strip()
-                explanation = ''
-                ticket_compliance_level = ''  # Individual ticket compliance
-                fully_compliant_str = ticket_analysis.get('fully_compliant_requirements', '').strip()
-                not_compliant_str = ticket_analysis.get('not_compliant_requirements', '').strip()
-                requires_further_human_verification = ticket_analysis.get('requires_further_human_verification',
-                                                                          '').strip()
-
-                if not fully_compliant_str and not not_compliant_str:
-                    get_logger().debug(f"Ticket compliance has no requirements",
-                                       artifact={'ticket_url': ticket_url})
-                    continue
-
-                # Calculate individual ticket compliance level
-                if fully_compliant_str:
-                    if not_compliant_str:
-                        ticket_compliance_level = 'Partially compliant'
-                    else:
-                        if not requires_further_human_verification:
-                            ticket_compliance_level = 'Fully compliant'
-                        else:
-                            ticket_compliance_level = 'PR Code Verified'
-                elif not_compliant_str:
-                    ticket_compliance_level = 'Not compliant'
-
-                # Store the compliance level for aggregation
-                if ticket_compliance_level:
-                    all_compliance_levels.append(ticket_compliance_level)
-
-                # build compliance string
-                if fully_compliant_str:
-                    explanation += f"Compliant requirements:\n\n{fully_compliant_str}\n\n"
-                if not_compliant_str:
-                    explanation += f"Non-compliant requirements:\n\n{not_compliant_str}\n\n"
-                if requires_further_human_verification:
-                    explanation += f"Requires further human verification:\n\n{requires_further_human_verification}\n\n"
-                ticket_compliance_str += f"\n\n**[{ticket_url.split('/')[-1]}]({ticket_url}) - {ticket_compliance_level}**\n\n{explanation}\n\n"
-
-                # for debugging
-                if requires_further_human_verification:
-                    get_logger().debug(f"Ticket compliance requires further human verification",
-                                       artifact={'ticket_url': ticket_url,
-                                                 'requires_further_human_verification': requires_further_human_verification,
-                                                 'compliance_level': ticket_compliance_level})
-
-            except Exception as e:
-                get_logger().exception(f"Failed to process ticket compliance: {e}")
-                continue
-
-        # Calculate overall compliance level and emoji
-        if all_compliance_levels:
-            if all(level == 'Fully compliant' for level in all_compliance_levels):
-                compliance_level = 'Fully compliant'
-                compliance_emoji = '✅'
-            elif all(level == 'PR Code Verified' for level in all_compliance_levels):
-                compliance_level = 'PR Code Verified'
-                compliance_emoji = '✅'
-            elif any(level == 'Not compliant' for level in all_compliance_levels):
-                # If there's a mix of compliant and non-compliant tickets
-                if any(level in ['Fully compliant', 'PR Code Verified'] for level in all_compliance_levels):
-                    compliance_level = 'Partially compliant'
-                    compliance_emoji = '🔶'
-                else:
-                    compliance_level = 'Not compliant'
-                    compliance_emoji = '❌'
-            elif any(level == 'Partially compliant' for level in all_compliance_levels):
-                compliance_level = 'Partially compliant'
-                compliance_emoji = '🔶'
-            else:
-                compliance_level = 'PR Code Verified'
-                compliance_emoji = '✅'
-
-            # Set extra statistics outside the ticket loop
-            get_settings().set('config.extra_statistics', {'compliance_level': compliance_level})
-
-        # editing table row for ticket compliance analysis
-        if gfm_supported:
-            markdown_text += f"<tr><td>\n\n"
-            markdown_text += f"**{emoji} Ticket compliance analysis {compliance_emoji}**\n\n"
-            markdown_text += ticket_compliance_str
-            markdown_text += f"</td></tr>\n"
-        else:
-            markdown_text += f"### {emoji} Ticket compliance analysis {compliance_emoji}\n\n"
-            markdown_text += ticket_compliance_str + "\n\n"
-
-    return markdown_text
-
-
-def process_can_be_split(emoji, value):
-    try:
-        # key_nice = "Can this PR be split?"
-        key_nice = "Multiple PR themes"
-        markdown_text = ""
-        if not value or isinstance(value, list) and len(value) == 1:
-            value = "No"
-            # markdown_text += f"<tr><td> {emoji}&nbsp;<strong>{key_nice}</strong></td><td>\n\n{value}\n\n</td></tr>\n"
-            # markdown_text += f"### {emoji} No multiple PR themes\n\n"
-            markdown_text += f"{emoji} <strong>No multiple PR themes</strong>\n\n"
-        else:
-            markdown_text += f"{emoji} <strong>{key_nice}</strong><br><br>\n\n"
-            for i, split in enumerate(value):
-                title = split.get('title', '')
-                relevant_files = split.get('relevant_files', [])
-                markdown_text += f"<details><summary>\nSub-PR theme: <b>{title}</b></summary>\n\n"
-                markdown_text += f"___\n\nRelevant files:\n\n"
-                for file in relevant_files:
-                    markdown_text += f"- {file}\n"
-                markdown_text += f"___\n\n"
-                markdown_text += f"</details>\n\n"
-
-                # markdown_text += f"#### Sub-PR theme: {title}\n\n"
-                # markdown_text += f"Relevant files:\n\n"
-                # for file in relevant_files:
-                #     markdown_text += f"- {file}\n"
-                # markdown_text += "\n"
-            # number_of_splits = len(value)
-            # markdown_text += f"<tr><td rowspan={number_of_splits}> {emoji}&nbsp;<strong>{key_nice}</strong></td>\n"
-            # for i, split in enumerate(value):
-            #     title = split.get('title', '')
-            #     relevant_files = split.get('relevant_files', [])
-            #     if i == 0:
-            #         markdown_text += f"<td><details><summary>\nSub-PR theme:<br><strong>{title}</strong></summary>\n\n"
-            #         markdown_text += f"<hr>\n"
-            #         markdown_text += f"Relevant files:\n"
-            #         markdown_text += f"<ul>\n"
-            #         for file in relevant_files:
-            #             markdown_text += f"<li>{file}</li>\n"
-            #         markdown_text += f"</ul>\n\n</details></td></tr>\n"
-            #     else:
-            #         markdown_text += f"<tr>\n<td><details><summary>\nSub-PR theme:<br><strong>{title}</strong></summary>\n\n"
-            #         markdown_text += f"<hr>\n"
-            #         markdown_text += f"Relevant files:\n"
-            #         markdown_text += f"<ul>\n"
-            #         for file in relevant_files:
-            #             markdown_text += f"<li>{file}</li>\n"
-            #         markdown_text += f"</ul>\n\n</details></td></tr>\n"
-    except Exception as e:
-        get_logger().exception(f"Failed to process can be split: {e}")
-        return ""
-    return markdown_text
-
-
-def parse_code_suggestion(code_suggestion: dict, i: int = 0, gfm_supported: bool = True) -> str:
-    """
-    Convert a dictionary of data into markdown format.
-
-    Args:
-        code_suggestion (dict): A dictionary containing data to be converted to markdown format.
-
-    Returns:
-        str: A string containing the markdown formatted text generated from the input dictionary.
-    """
-    markdown_text = ""
-    if gfm_supported and 'relevant_line' in code_suggestion:
-        markdown_text += '<table>'
-        for sub_key, sub_value in code_suggestion.items():
-            try:
-                if sub_key.lower() == 'relevant_file':
-                    relevant_file = sub_value.strip('`').strip('"').strip("'")
-                    markdown_text += f"<tr><td>relevant file</td><td>{relevant_file}</td></tr>"
-                    # continue
-                elif sub_key.lower() == 'suggestion':
-                    markdown_text += (f"<tr><td>{sub_key} &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</td>"
-                                      f"<td>\n\n<strong>\n\n{sub_value.strip()}\n\n</strong>\n</td></tr>")
-                elif sub_key.lower() == 'relevant_line':
-                    markdown_text += f"<tr><td>relevant line</td>"
-                    sub_value_list = sub_value.split('](')
-                    relevant_line = sub_value_list[0].lstrip('`').lstrip('[')
-                    if len(sub_value_list) > 1:
-                        link = sub_value_list[1].rstrip(')').strip('`')
-                        markdown_text += f"<td><a href='{link}'>{relevant_line}</a></td>"
-                    else:
-                        markdown_text += f"<td>{relevant_line}</td>"
-                    markdown_text += "</tr>"
-            except Exception as e:
-                get_logger().exception(f"Failed to parse code suggestion: {e}")
-                pass
-        markdown_text += '</table>'
-        markdown_text += "<hr>"
-    else:
-        for sub_key, sub_value in code_suggestion.items():
-            if isinstance(sub_key, str):
-                sub_key = sub_key.rstrip()
-            if isinstance(sub_value,str):
-                sub_value = sub_value.rstrip()
-            if isinstance(sub_value, dict):  # "code example"
-                markdown_text += f"  - **{sub_key}:**\n"
-                for code_key, code_value in sub_value.items():  # 'before' and 'after' code
-                    code_str = f"```\n{code_value}\n```"
-                    code_str_indented = textwrap.indent(code_str, '        ')
-                    markdown_text += f"    - **{code_key}:**\n{code_str_indented}\n"
-            else:
-                if "relevant_file" in sub_key.lower():
-                    markdown_text += f"\n  - **{sub_key}:** {sub_value}  \n"
-                else:
-                    markdown_text += f"   **{sub_key}:** {sub_value}  \n"
-                if "relevant_line" not in sub_key.lower():  # nicer presentation
-                    # markdown_text = markdown_text.rstrip('\n') + "\\\n" # works for gitlab
-                    markdown_text = markdown_text.rstrip('\n') + "   \n"  # works for gitlab and bitbucker
-
-        markdown_text += "\n"
-    return markdown_text
-
-
-def try_fix_json(review, max_iter=10, code_suggestions=False):
-    """
-    Fix broken or incomplete JSON messages and return the parsed JSON data.
-
-    Args:
-    - review: A string containing the JSON message to be fixed.
-    - max_iter: An integer representing the maximum number of iterations to try and fix the JSON message.
-    - code_suggestions: A boolean indicating whether to try and fix JSON messages with code feedback.
-
-    Returns:
-    - data: A dictionary containing the parsed JSON data.
-
-    The function attempts to fix broken or incomplete JSON messages by parsing until the last valid code suggestion.
-    If the JSON message ends with a closing bracket, the function calls the fix_json_escape_char function to fix the
-    message.
-    If code_suggestions is True and the JSON message contains code feedback, the function tries to fix the JSON
-    message by parsing until the last valid code suggestion.
-    The function uses regular expressions to find the last occurrence of "}," with any number of whitespaces or
-    newlines.
-    It tries to parse the JSON message with the closing bracket and checks if it is valid.
-    If the JSON message is valid, the parsed JSON data is returned.
-    If the JSON message is not valid, the last code suggestion is removed and the process is repeated until a valid JSON
-    message is obtained or the maximum number of iterations is reached.
-    If a valid JSON message is not obtained, an error is logged and an empty dictionary is returned.
-    """
-
-    if review.endswith("}"):
-        return fix_json_escape_char(review)
-
-    data = {}
-    if code_suggestions:
-        closing_bracket = "]}"
-    else:
-        closing_bracket = "]}}"
-
-    if (review.rfind("'Code feedback': [") > 0 or review.rfind('"Code feedback": [') > 0) or \
-            (review.rfind("'Code suggestions': [") > 0 or review.rfind('"Code suggestions": [') > 0) :
-        last_code_suggestion_ind = [m.end() for m in re.finditer(r"\}\s*,", review)][-1] - 1
-        valid_json = False
-        iter_count = 0
-
-        while last_code_suggestion_ind > 0 and not valid_json and iter_count < max_iter:
-            try:
-                data = json.loads(review[:last_code_suggestion_ind] + closing_bracket)
-                valid_json = True
-                review = review[:last_code_suggestion_ind].strip() + closing_bracket
-            except json.decoder.JSONDecodeError:
-                review = review[:last_code_suggestion_ind]
-                last_code_suggestion_ind = [m.end() for m in re.finditer(r"\}\s*,", review)][-1] - 1
-                iter_count += 1
-
-        if not valid_json:
-            get_logger().error("Unable to decode JSON response from AI")
-            data = {}
-
-    return data
-
-
-def fix_json_escape_char(json_message=None):
-    """
-    Fix broken or incomplete JSON messages and return the parsed JSON data.
-
-    Args:
-        json_message (str): A string containing the JSON message to be fixed.
-
-    Returns:
-        dict: A dictionary containing the parsed JSON data.
-
-    Raises:
-        None
-
-    """
-    try:
-        result = json.loads(json_message)
-    except Exception as e:
-        # Find the offending character index:
-        idx_to_replace = int(str(e).split(' ')[-1].replace(')', ''))
-        # Remove the offending character:
-        json_message = list(json_message)
-        json_message[idx_to_replace] = ' '
-        new_message = ''.join(json_message)
-        return fix_json_escape_char(json_message=new_message)
-    return result
 
 
 def convert_str_to_datetime(date_str):
@@ -747,196 +146,6 @@ def _fix_key_value(key: str, value: str):
     except Exception as e:
         get_logger().debug(f"Failed to parse YAML for config override {key}={value}", exc_info=e)
     return key, value
-
-
-def load_yaml(response_text: str, keys_fix_yaml: List[str] = [], first_key="", last_key="") -> dict:
-    response_text_original = copy.deepcopy(response_text)
-    response_text = response_text.strip('\n').removeprefix('yaml').removeprefix('```yaml').rstrip().removesuffix('```')
-    try:
-        data = yaml.safe_load(response_text)
-    except Exception as e:
-        get_logger().warning(f"Initial failure to parse AI prediction: {e}")
-        data = try_fix_yaml(response_text, keys_fix_yaml=keys_fix_yaml, first_key=first_key, last_key=last_key,
-                            response_text_original=response_text_original)
-        if not data:
-            get_logger().error(f"Failed to parse AI prediction after fallbacks",
-                               artifact={'response_text': response_text})
-        else:
-            get_logger().info(f"Successfully parsed AI prediction after fallbacks",
-                              artifact={'response_text': response_text})
-    return data
-
-
-
-def try_fix_yaml(response_text: str,
-                 keys_fix_yaml: List[str] = [],
-                 first_key="",
-                 last_key="",
-                 response_text_original="") -> dict:
-    response_text_lines = response_text.split('\n')
-
-    keys_yaml = ['relevant line:', 'suggestion content:', 'relevant file:', 'existing code:',
-                 'improved code:', 'label:', 'why:', 'suggestion_summary:']
-    keys_yaml = keys_yaml + keys_fix_yaml
-
-    # first fallback - try to convert 'relevant line: ...' to relevant line: |-\n        ...'
-    response_text_lines_copy = response_text_lines.copy()
-    for i in range(0, len(response_text_lines_copy)):
-        for key in keys_yaml:
-            if key in response_text_lines_copy[i] and not '|' in response_text_lines_copy[i]:
-                response_text_lines_copy[i] = response_text_lines_copy[i].replace(f'{key}',
-                                                                                  f'{key} |\n        ')
-    try:
-        data = yaml.safe_load('\n'.join(response_text_lines_copy))
-        get_logger().info(f"Successfully parsed AI prediction after adding |-\n")
-        return data
-    except:
-        pass
-
-    # 1.5 fallback - try to convert '|' to '|2'. Will solve cases of indent decreasing during the code
-    response_text_copy = copy.deepcopy(response_text)
-    response_text_copy = response_text_copy.replace('|\n', '|2\n')
-    try:
-        data = yaml.safe_load(response_text_copy)
-        get_logger().info(f"Successfully parsed AI prediction after replacing | with |2")
-        return data
-    except:
-        # if it fails, we can try to add spaces to the lines that are not indented properly, and contain '}'.
-        response_text_lines_copy = response_text_copy.split('\n')
-        for i in range(0, len(response_text_lines_copy)):
-            initial_space = len(response_text_lines_copy[i]) - len(response_text_lines_copy[i].lstrip())
-            if initial_space == 2 and '|2' not in response_text_lines_copy[i] and '}' in response_text_lines_copy[i]:
-                response_text_lines_copy[i] = '    ' + response_text_lines_copy[i].lstrip()
-        try:
-            data = yaml.safe_load('\n'.join(response_text_lines_copy))
-            get_logger().info(f"Successfully parsed AI prediction after replacing | with |2 and adding spaces")
-            return data
-        except:
-            pass
-
-    # second fallback - try to extract only range from first ```yaml to the last ```
-    snippet_pattern = r'```(yaml|yml)?([\s\S]*?)```(?=\s*$|")'
-    snippet = re.search(snippet_pattern, '\n'.join(response_text_lines_copy))
-    if not snippet:
-        snippet = re.search(snippet_pattern, response_text_original) # before we removed the "```"
-    if snippet:
-        # group(2) is the snippet body, without the ``` fences or the optional yaml/yml language identifier
-        snippet_text = snippet.group(2)
-        try:
-            data = yaml.safe_load(snippet_text)
-            get_logger().info(f"Successfully parsed AI prediction after extracting yaml snippet")
-            return data
-        except Exception as e:
-            get_logger().debug(f"Failed to parse AI prediction after extracting yaml snippet: {e}")
-
-
-    # third fallback - try to remove leading and trailing curly brackets
-    response_text_copy = response_text.strip().rstrip().removeprefix('{').removesuffix('}').rstrip(':\n')
-    try:
-        data = yaml.safe_load(response_text_copy)
-        get_logger().info(f"Successfully parsed AI prediction after removing curly brackets")
-        return data
-    except:
-        pass
-
-
-    # forth fallback - try to extract yaml snippet by 'first_key' and 'last_key'
-    # note that 'last_key' can be in practice a key that is not the last key in the yaml snippet.
-    # it just needs to be some inner key, so we can look for newlines after it
-    if first_key and last_key:
-        index_start = response_text.find(f"\n{first_key}:")
-        if index_start == -1:
-            index_start = response_text.find(f"{first_key}:")
-        index_last_code = response_text.rfind(f"{last_key}:")
-        index_end = response_text.find("\n\n", index_last_code) # look for newlines after last_key
-        if index_end == -1:
-            index_end = len(response_text)
-        response_text_copy = response_text[index_start:index_end].strip().strip('```yaml').strip('`').strip()
-        if response_text_copy:
-            try:
-                data = yaml.safe_load(response_text_copy)
-                get_logger().info(f"Successfully parsed AI prediction after extracting yaml snippet")
-                return data
-            except:
-                pass
-
-    # fifth fallback - try to remove leading '+' (sometimes added by AI for 'existing code' and 'improved code')
-    response_text_lines_copy = response_text_lines.copy()
-    for i in range(0, len(response_text_lines_copy)):
-        if response_text_lines_copy[i].startswith('+'):
-            response_text_lines_copy[i] = ' ' + response_text_lines_copy[i][1:]
-    try:
-        data = yaml.safe_load('\n'.join(response_text_lines_copy))
-        get_logger().info(f"Successfully parsed AI prediction after removing leading '+'")
-        return data
-    except:
-        pass
-
-    # sixth fallback - replace tabs with spaces
-    if '\t' in response_text:
-        response_text_copy = copy.deepcopy(response_text)
-        response_text_copy = response_text_copy.replace('\t', '    ')
-        try:
-            data = yaml.safe_load(response_text_copy)
-            get_logger().info(f"Successfully parsed AI prediction after replacing tabs with spaces")
-            return data
-        except:
-            pass
-
-    # seventh fallback - add indent for sections of code blocks
-    response_text_copy = copy.deepcopy(response_text)
-    response_text_copy_lines = response_text_copy.split('\n')
-    start_line = -1
-    improve_sections = ['existing_code:', 'improved_code:', 'response:', 'why:']
-    describe_sections = ['description:', 'title:', 'changes_diagram:', 'pr_files:', 'pr_ticket:']
-    for i, line in enumerate(response_text_copy_lines):
-        line_stripped = line.rstrip()
-        if any(key in line_stripped for key in (improve_sections+describe_sections)):
-            start_line = i
-        elif line_stripped.endswith(': |') or line_stripped.endswith(': |-') or line_stripped.endswith(': |2') or any(line_stripped.endswith(key) for key in keys_yaml):
-            start_line = -1
-        elif start_line != -1:
-            response_text_copy_lines[i] = '    ' + line
-    response_text_copy = '\n'.join(response_text_copy_lines)
-    response_text_copy = response_text_copy.replace(' |\n', ' |2\n')
-    try:
-        data = yaml.safe_load(response_text_copy)
-        get_logger().info(f"Successfully parsed AI prediction after adding indent for sections of code blocks")
-        return data
-    except:
-        pass
-
-    # eighth fallback - try to remove pipe chars at the root-level dicts
-    response_text_copy = copy.deepcopy(response_text)
-    response_text_copy = response_text_copy.lstrip('|\n')
-    try:
-        data = yaml.safe_load(response_text_copy)
-        get_logger().info(f"Successfully parsed AI prediction after removing pipe chars")
-        return data
-    except:
-        pass
-
-    # ninth fallback - try to decode the response text with different encodings. GPT-5 can return text that is not utf-8 encoded.
-    encodings_to_try = ['latin-1', 'utf-16']
-    for encoding in encodings_to_try:
-        try:
-            data = yaml.safe_load(response_text.encode(encoding).decode("utf-8"))
-            if data:
-                get_logger().info(f"Successfully parsed AI prediction after decoding with {encoding} encoding")
-                return data
-        except:
-            pass
-
-    # # sixth fallback - try to remove last lines
-    # for i in range(1, len(response_text_lines)):
-    #     response_text_lines_tmp = '\n'.join(response_text_lines[:-i])
-    #     try:
-    #         data = yaml.safe_load(response_text_lines_tmp)
-    #         get_logger().info(f"Successfully parsed AI prediction after removing {i} lines")
-    #         return data
-    #     except:
-    #         pass
-
 
 
 def set_custom_labels(variables, git_provider=None):
@@ -1132,17 +341,6 @@ def clip_tokens(text: str, max_tokens: int, add_three_dots=True, num_input_token
         get_logger().warning(f"Failed to clip tokens: {e}")
         return text
 
-def replace_code_tags(text):
-    """
-    Replace odd instances of ` with <code> and even instances of ` with </code>
-    """
-    text = html.escape(text)
-    parts = text.split('`')
-    for i in range(1, len(parts), 2):
-        parts[i] = '<code>' + parts[i] + '</code>'
-    return ''.join(parts)
-
-
 def find_line_number_of_relevant_line_in_file(diff_files: List[FilePatchInfo],
                                               relevant_file: str,
                                               relevant_line_in_file: str,
@@ -1317,15 +515,6 @@ def show_relevant_configurations(relevant_section: str) -> str:
     markdown_text += "\n</details>\n"
     return markdown_text
 
-def is_value_no(value):
-    if not value:
-        return True
-    value_str = str(value).strip().lower()
-    if value_str == 'no' or value_str == 'none' or value_str == 'false':
-        return True
-    return False
-
-
 def set_pr_string(repo_name, pr_number):
     return f"{repo_name}#{pr_number}"
 
@@ -1487,46 +676,3 @@ def set_file_languages(diff_files) -> List[FilePatchInfo]:
 
     return diff_files
 
-def format_todo_item(todo_item: TodoItem, git_provider, gfm_supported) -> str:
-    relevant_file = todo_item.get('relevant_file', '').strip()
-    line_number = todo_item.get('line_number', '')
-    content = todo_item.get('content', '')
-    reference_link = git_provider.get_line_link(relevant_file, line_number, line_number)
-    file_ref = f"{relevant_file} [{line_number}]"
-    if reference_link:
-        if gfm_supported:
-            file_ref = f"<a href='{reference_link}'>{file_ref}</a>"
-        else:
-            file_ref = f"[{file_ref}]({reference_link})"
-
-    if content:
-        return f"{file_ref}: {content.strip()}"
-    else:
-        # if content is empty, return only the file reference
-        return file_ref
-
-
-def format_todo_items(value: list[TodoItem] | TodoItem, git_provider, gfm_supported) -> str:
-    markdown_text = ""
-    MAX_ITEMS = 5 # limit the number of items to display
-    if gfm_supported:
-        if isinstance(value, list):
-            markdown_text += "<ul>\n"
-            if len(value) > MAX_ITEMS:
-                get_logger().debug(f"Truncating todo items to {MAX_ITEMS} items")
-                value = value[:MAX_ITEMS]
-            for todo_item in value:
-                markdown_text += f"<li>{format_todo_item(todo_item, git_provider, gfm_supported)}</li>\n"
-            markdown_text += "</ul>\n"
-        else:
-            markdown_text += f"<p>{format_todo_item(value, git_provider, gfm_supported)}</p>\n"
-    else:
-        if isinstance(value, list):
-            if len(value) > MAX_ITEMS:
-                get_logger().debug(f"Truncating todo items to {MAX_ITEMS} items")
-                value = value[:MAX_ITEMS]
-            for todo_item in value:
-                markdown_text += f"- {format_todo_item(todo_item, git_provider, gfm_supported)}\n"
-        else:
-            markdown_text += f"- {format_todo_item(value, git_provider, gfm_supported)}\n"
-    return markdown_text
